CI 수리 — depth_pct INT 정렬·비콘 DTO를 presentation/request 로 분리
Hibernate validate 가 SMALLINT 를 Integer 와 불일치로 거부(CI 첫 실패). 소스 규칙(컨트롤러 인라인 DTO 금지·Request 파일 위치)도 함께 준수. 로컬 전수 테스트는 Redis 6380·서브모듈 초기화 후 그린.

+++ src/main/resources/db/migration/V115__behavior_event.sql
@@ -10,7 +10,7 @@ CREATE TABLE behavior_event (
     post_id      BIGINT      NULL,
     target_type  VARCHAR(32) NULL,
     target_id    VARCHAR(64) NULL,
-    depth_pct    SMALLINT    NULL,
+    depth_pct    INT         NULL,
     dwell_ms     BIGINT      NULL,
     device_class VARCHAR(32) NULL,
     is_bot       BOOLEAN     NOT NULL DEFAULT FALSE,
